requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.2
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
bleach==6.2.0
certifi==2025.1.31
cffi==2.0.0
channels==4.3.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
cycler==0.12.1
daphne==4.2.1
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.1.6
django-allauth==65.7.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.19.1
flatbuffers==25.1.24
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.9.0
gast==0.6.0
geographiclib==2.1
geopy==2.4.1
git-filter-repo==2.47.0
gotrue==2.11.4
grpcio==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pika==1.3.2
pillow==11.1.0
platformdirs==4.3.6
postgrest==0.19.3
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.2.1
python-pptx==1.0.2
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
realtime==2.4.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
Send2Trash==1.8.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.14.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.9.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.15.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xlsxwriter==3.2.9
yarl==1.18.3
zope.interface==8.0.1
# Core Django
Django==5.1.6
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.10  # Use binary version for easier install
dj-database-url==2.3.0

# Authentication (if needed)
django-allauth==65.7.0
PyJWT==2.10.1

# CORS (if needed)
django-cors-headers==4.9.0

# Async/Channels (if needed)
channels==4.3.1
daphne==4.2.1
twisted==25.5.0

# File handling
Pillow==11.1.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.1

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# Common utilities
requests==2.32.3

# Core Django
Django==5.1.6
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# File handling
Pillow==11.1.0

# Environment variables
python-decouple==3.8